Quick fix to import loop between constants and util
